Integrate: [ 25211] Silence Win32 warnings when not using USE_LARGE_FILES p4raw-link: @25211 on //depot/perl: 8859a7a0b5d2e549800ebdce5712e285a2861e3d
p4raw-id: //depot/maint-5.8/perl@25667
p4raw-integrated: from //depot/perl@25211 'merge in' win32/win32.c
	(@25101..)

+++ win32/win32.c
@@ -2642,7 +2642,7 @@ win32_fseek(FILE *pf, Off_t offset,int origin)
     return fsetpos(pf, &offset);
 #endif
 #else
-    return fseek(pf, offset, origin);
+    return fseek(pf, (long)offset, origin);
 #endif
 }
 
@@ -3192,7 +3192,7 @@ finish:
     win32_lseek(fd, cur, SEEK_SET);
     return retval;
 #else
-    return chsize(fd, size);
+    return chsize(fd, (long)size);
 #endif
 }
 
@@ -3218,7 +3218,7 @@ win32_lseek(int fd, Off_t offset, int origin)
     return _lseeki64(fd, offset, origin);
 #endif
 #else
-    return lseek(fd, offset, origin);
+    return lseek(fd, (long)offset, origin);
 #endif
 }
